requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.32.0
groq>=0.5.0
google-generativeai>=0.4.0
openai>=1.10.0